Training a Forecasting Model with Linear Regression Framework

Methods for forecasting regular time series data based on a linear regression model

trainLM(input, y = NULL, x = NULL, seasonal = NULL,
  trend = list(linear = TRUE, exponential = FALSE, log = FALSE, power =
  FALSE), lags = NULL, splines = NULL, events = NULL, scale = NULL,
  step = FALSE, ...)

Arguments

input
A tsibble or ts object

y
A character, the column name of the depended variable of the input object, required (and applicable) only when the input is tsibble object

x
A character, the column names of the independent variable of the input object, applicable when using tsibble object with regressors

seasonal
A character, optional, create categorical variable/s to model the single or multiple seasonal components. Supporting the following frequencies structure:

`quarter` - for quarterly frequency

`month` - for monthly frequency

`week` - for weekly frequency

`yday` - for daily frequency

`wday` - for day of the week frequency

`hour` - for hourly frequency

The argument supports multiple seasonality cases such as daily and day of the week effect

trend
A list, define the trend structure. Possible arguments -

`linear`, a boolean variable, if set to TRUE defines a linear trend (e.g., index of 1,2,3,...,t, for a series with t observations)

`power``, a numeric value, defines the polynomial degree of the series index (for example, a power = 0.5 define a square root index and power = 2 represents a squared index). By default set to NULL

`exponential`, a boolean variable, if set to TRUE defines an exponential trend

`log`, a boolean variable, if set to TRUE defines a log transformation for the trend.

By default, the `trend` argument is set to a linear trend (i.e., power = 1)

lags
A positive integer, defines the series lags to be used as input to the model (equivalent to AR process)

splines
A list, optional, provides the ability to model structural shifts or breaks of the series trend with the use of spline or piecewise regression. The argument enables to define either single or multiple splines /piecewise regression by setting for each one as a list with the following two arguments:

type - defines the spline feature. For model a linear trend for each segments, set the type argument as `linear`. For a structural break or shift of the trend, set the type argument as `break`.

knots - a single or sequence of dates/time that represents the breaking point of the trend (either slope shift or structural break). Note when using `tsibble`` object as input, the dates or time inputs must align with the series index class. When using `ts` object, the input class should align with the converted index of the `ts`` object. For example, if the frequency of the `ts` object is 12 (i.e., monthly), you should use `Date`` objects to set the knots

events
A list, optional, create hot encoding variables based on date/time objects, where the date/time objects must align with the input object index class (may not work when the input object is 'ts')

scale
A character, scaling options of the series, methods available - c("log", "normal", "standard") for log transformation, normalization, or standardization of the series, respectively. If set to NULL (default), no transformation will occur

step
A boolean, if set to TRUE will use apply the stepwise function for variable selection using the step function

...
Optional, a list of arguments to pass to the step function

Details

The trainLM function provides a flexible framework for forecasting regular time-series data with the linear regression model with the lm function. The function arguments enable a fast extraction and generation of new features from the input series, such as seasonal, trend, lags, outliers, and special events.

Examples

data(ny_gas)

head(ny_gas)

#> # A tsibble: 6 x 2 [1M]
#>       date     y
#>      <mth> <int>
#> 1 1997 Jan 62019
#> 2 1997 Feb 56003
#> 3 1997 Mar 51213
#> 4 1997 Apr 37597
#> 5 1997 May 24082
#> 6 1997 Jun 14469

# Fitting basic model with linear trend
md1 <- trainLM(input = ny_gas,
               y = "y",
               trend = list(linear = TRUE))

# Getting the regression summary

summary(md1$model)

#> 
#> Call:
#> stats::lm(formula = f, data = df1)
#> 
#> Residuals:
#>    Min     1Q Median     3Q    Max 
#> -27370 -21542  -6345  22692  55294 
#> 
#> Coefficients:
#>              Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  30101.85    2808.10  10.720   <2e-16 ***
#> linear_trend    26.20      17.51   1.496    0.136    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 23300 on 275 degrees of freedom
#> Multiple R-squared:  0.008074,	Adjusted R-squared:  0.004467 
#> F-statistic: 2.238 on 1 and 275 DF,  p-value: 0.1358
#> 

# Plotting the residauls of the model
plot_res(md1)

# Adding monthly seasonal component
md2 <- trainLM(input = ny_gas,
               y = "y",
               trend = list(linear = TRUE),
               seasonal = "month")

plot_res(md2)

# Adding the first and seasonal lags

md3 <- trainLM(input = ny_gas,
               y = "y",
               trend = list(linear = TRUE),
               seasonal = "month",
               lags = c(1, 12))

plot_res(md3)

# Adding more lags and using stepwise regression for variable selection
md4 <- trainLM(input = ny_gas,
               y = "y",
               trend = list(linear = TRUE),
               seasonal = "month",
               lags = c(1:12),
               step = TRUE)

#> Start:  AIC=4528.57
#> y ~ month + linear_trend + lag_1 + lag_2 + lag_3 + lag_4 + lag_5 + 
#>     lag_6 + lag_7 + lag_8 + lag_9 + lag_10 + lag_11 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_2         1     337090 5793758744 4526.6
#> - lag_4         1    3105556 5796527210 4526.7
#> - lag_8         1    4798094 5798219748 4526.8
#> - lag_7         1    9093498 5802515152 4527.0
#> - lag_11        1   12684502 5806106156 4527.1
#> - lag_6         1   12938759 5806360413 4527.2
#> - lag_12        1   21024489 5814446143 4527.5
#> - lag_3         1   33099600 5826521254 4528.1
#> <none>                       5793421654 4528.6
#> - lag_9         1   45100935 5838522589 4528.6
#> - lag_10        1   85876052 5879297706 4530.5
#> - lag_5         1  119772165 5913193819 4532.0
#> - linear_trend  1  140768290 5934189944 4532.9
#> - lag_1         1 1377572660 7170994314 4583.1
#> - month        11 2042112108 7835533763 4586.6
#> 
#> Step:  AIC=4526.58
#> y ~ month + linear_trend + lag_1 + lag_3 + lag_4 + lag_5 + lag_6 + 
#>     lag_7 + lag_8 + lag_9 + lag_10 + lag_11 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_4         1    3087682 5796846426 4524.7
#> - lag_8         1    4662499 5798421243 4524.8
#> - lag_7         1    8776507 5802535251 4525.0
#> - lag_6         1   12807126 5806565870 4525.2
#> - lag_11        1   13716246 5807474990 4525.2
#> - lag_12        1   20696514 5814455258 4525.5
#> - lag_3         1   36232633 5829991377 4526.2
#> <none>                       5793758744 4526.6
#> - lag_9         1   44793857 5838552601 4526.6
#> - lag_10        1   85563041 5879321785 4528.5
#> - lag_5         1  121071439 5914830183 4530.1
#> - linear_trend  1  140614667 5934373411 4530.9
#> - month        11 2079203070 7872961814 4585.8
#> - lag_1         1 1703541785 7497300529 4592.9
#> 
#> Step:  AIC=4524.73
#> y ~ month + linear_trend + lag_1 + lag_3 + lag_5 + lag_6 + lag_7 + 
#>     lag_8 + lag_9 + lag_10 + lag_11 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_8         1    4244935 5801091361 4522.9
#> - lag_7         1    9376841 5806223267 4523.2
#> - lag_6         1   12710637 5809557063 4523.3
#> - lag_11        1   14068483 5810914909 4523.4
#> - lag_12        1   20688777 5817535203 4523.7
#> - lag_3         1   33963818 5830810244 4524.3
#> - lag_9         1   42172556 5839018982 4524.6
#> <none>                       5796846426 4524.7
#> - lag_10        1   84400560 5881246986 4526.6
#> - linear_trend  1  138308669 5935155095 4529.0
#> - lag_5         1  161848690 5958695116 4530.0
#> - month        11 2105368658 7902215084 4584.8
#> - lag_1         1 1700796960 7497643386 4590.9
#> 
#> Step:  AIC=4522.92
#> y ~ month + linear_trend + lag_1 + lag_3 + lag_5 + lag_6 + lag_7 + 
#>     lag_9 + lag_10 + lag_11 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_7         1    6029808 5807121168 4521.2
#> - lag_6         1   12830119 5813921480 4521.5
#> - lag_11        1   14983446 5816074807 4521.6
#> - lag_12        1   20345128 5821436489 4521.8
#> - lag_3         1   31240855 5832332216 4522.3
#> - lag_9         1   38217452 5839308812 4522.7
#> <none>                       5801091361 4522.9
#> - lag_10        1   84231235 5885322596 4524.7
#> - linear_trend  1  144099484 5945190845 4527.4
#> - lag_5         1  158555665 5959647026 4528.1
#> - month        11 2159392317 7960483678 4584.8
#> - lag_1         1 1697649607 7498740968 4588.9
#> 
#> Step:  AIC=4521.19
#> y ~ month + linear_trend + lag_1 + lag_3 + lag_5 + lag_6 + lag_9 + 
#>     lag_10 + lag_11 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_6         1    7873732 5814994900 4519.6
#> - lag_11        1   13452358 5820573526 4519.8
#> - lag_12        1   25529986 5832651154 4520.4
#> - lag_3         1   32891889 5840013057 4520.7
#> <none>                       5807121168 4521.2
#> - lag_9         1   45566962 5852688131 4521.3
#> - lag_10        1   81980878 5889102047 4522.9
#> - linear_trend  1  138705952 5945827120 4525.4
#> - lag_5         1  163064609 5970185778 4526.5
#> - month        11 2198506428 8005627597 4584.3
#> - lag_1         1 1703838083 7510959252 4587.4
#> 
#> Step:  AIC=4519.55
#> y ~ month + linear_trend + lag_1 + lag_3 + lag_5 + lag_9 + lag_10 + 
#>     lag_11 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_11        1   11169433 5826164333 4518.1
#> - lag_12        1   24854886 5839849786 4518.7
#> - lag_3         1   35573319 5850568219 4519.2
#> - lag_9         1   40881144 5855876044 4519.4
#> <none>                       5814994900 4519.6
#> - lag_10        1   86725765 5901720665 4521.5
#> - linear_trend  1  156592675 5971587575 4524.6
#> - lag_5         1  172726068 5987720968 4525.3
#> - month        11 2208989904 8023984804 4582.9
#> - lag_1         1 1713215586 7528210486 4586.0
#> 
#> Step:  AIC=4518.06
#> y ~ month + linear_trend + lag_1 + lag_3 + lag_5 + lag_9 + lag_10 + 
#>     lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_3         1   33832858 5859997191 4517.6
#> - lag_9         1   40288137 5866452469 4517.9
#> <none>                       5826164333 4518.1
#> - lag_12        1   52449956 5878614289 4518.4
#> - lag_10        1  138343890 5964508222 4522.3
#> - linear_trend  1  161829849 5987994182 4523.3
#> - lag_5         1  175113462 6001277795 4523.9
#> - lag_1         1 1758580984 7584745317 4586.0
#> - month        11 2474223579 8300387912 4589.9
#> 
#> Step:  AIC=4517.6
#> y ~ month + linear_trend + lag_1 + lag_5 + lag_9 + lag_10 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> <none>                       5859997191 4517.6
#> - lag_12        1   46792286 5906789477 4517.7
#> - lag_9         1   49405301 5909402492 4517.8
#> - lag_10        1  138424887 5998422078 4521.8
#> - lag_5         1  145037965 6005035156 4522.1
#> - linear_trend  1  198820490 6058817681 4524.4
#> - lag_1         1 2083214939 7943212130 4596.2
#> - month        11 3145083770 9005080961 4609.5

summary(md4$model)

#> 
#> Call:
#> stats::lm(formula = y ~ month + linear_trend + lag_1 + lag_5 + 
#>     lag_9 + lag_10 + lag_12, data = df1)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -15814.4  -2199.0    -26.3   2223.1  15024.0 
#> 
#> Coefficients:
#>                Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   2.987e+04  4.815e+03   6.204 2.30e-09 ***
#> monthFeb     -9.628e+03  1.976e+03  -4.872 1.97e-06 ***
#> monthMar     -1.137e+04  2.478e+03  -4.587 7.16e-06 ***
#> monthApr     -2.429e+04  3.141e+03  -7.732 2.67e-13 ***
#> monthMay     -2.453e+04  4.052e+03  -6.055 5.19e-09 ***
#> monthJun     -2.167e+04  4.843e+03  -4.475 1.17e-05 ***
#> monthJul     -2.024e+04  4.869e+03  -4.157 4.45e-05 ***
#> monthAug     -2.029e+04  4.721e+03  -4.298 2.48e-05 ***
#> monthSep     -2.320e+04  4.412e+03  -5.258 3.16e-07 ***
#> monthOct     -2.061e+04  4.217e+03  -4.886 1.85e-06 ***
#> monthNov     -1.090e+04  3.328e+03  -3.276  0.00120 ** 
#> monthDec     -1.585e+03  2.210e+03  -0.717  0.47399    
#> linear_trend  1.444e+01  4.987e+00   2.895  0.00413 ** 
#> lag_1         5.042e-01  5.381e-02   9.371  < 2e-16 ***
#> lag_5        -1.339e-01  5.417e-02  -2.473  0.01409 *  
#> lag_9        -9.074e-02  6.288e-02  -1.443  0.15027    
#> lag_10        1.583e-01  6.556e-02   2.415  0.01644 *  
#> lag_12        8.055e-02  5.736e-02   1.404  0.16146    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 4871 on 247 degrees of freedom
#> Multiple R-squared:  0.9599,	Adjusted R-squared:  0.9572 
#> F-statistic: 348.3 on 17 and 247 DF,  p-value: < 2.2e-16
#> 

plot_res(md4)