Training a Forecasting Model with Linear Regression Framework

Methods for forecasting regular time series data based on a linear regression model

trainLM(input, y = NULL, x = NULL, seasonal = NULL,
  trend = list(linear = TRUE, exponential = FALSE, log = FALSE, power =
  FALSE), lags = NULL, knots = NULL, events = NULL, scale = NULL,
  step = FALSE, ...)

Arguments

input
A tsibble or ts object

y
A character, the column name of the depended variable of the input object, required (and applicable) only when the input is tsibble object

x
A character, the column names of the independent variable of the input object, applicable when using tsibble object with regressors

seasonal
A character, optional, create categorical variable/s to model the single or multiple seasonal components. Supporting the following frequencies structure:

`quarter` - for quarterly frequency

`month` - for monthly frequency

`week` - for weekly frequency

`yday` - for daily frequency

`wday` - for day of the week frequency

`hour` - for hourly frequency

The argument supports multiple seasonality cases such as daily and day of the week effect

trend
A list, define the trend structure. Possible arguments -

`linear`, a boolean variable, if set to TRUE defines a linear trend (e.g., index of 1,2,3,...,t, for a series with t observations)

`power``, a numeric value, defines the polynomial degree of the series index (for example, a power = 0.5 define a square root index and power = 2 represents a squared index). By default set to NULL

`exponential`, a boolean variable, if set to TRUE defines an exponential trend

`log`, a boolean variable, if set to TRUE defines a log transformation for the trend.

By default, the `trend` argument is set to a linear trend (i.e., power = 1)

lags
A positive integer, defines the series lags to be used as input to the model (equivalent to AR process)

knots
A list, optional, create a piecewise linear trend variables based on date/time objects as a starting point of each knot, where the date/time objects must align with the input object index class (may not work when the input object is 'ts')

events
A list, optional, create hot encoding variables based on date/time objects, where the date/time objects must align with the input object index class (may not work when the input object is 'ts')

scale
A character, scaling options of the series, methods available - c("log", "normal", "standard") for log transformation, normalization, or standardization of the series, respectively. If set to NULL (default), no transformation will occur

step
A boolean, if set to TRUE will use apply the stepwise function for variable selection using the step function

...
Optional, a list of arguments to pass to the step function

Details

The trainLM function provides a flexible framework for forecasting regular time-series data with the linear regression model with the lm function. The function arguments enable a fast extraction and generation of new features from the input series, such as seasonal, trend, lags, outliers, and special events.

Examples

data(ny_gas)

head(ny_gas)

#> # A tsibble: 6 x 2 [1M]
#>       date     y
#>      <mth> <int>
#> 1 1997 Jan 62019
#> 2 1997 Feb 56003
#> 3 1997 Mar 51213
#> 4 1997 Apr 37597
#> 5 1997 May 24082
#> 6 1997 Jun 14469

# Fitting basic model with linear trend
md1 <- trainLM(input = ny_gas,
               y = "y",
               trend = list(linear = TRUE))

# Getting the regression summary

summary(md1$model)

#> 
#> Call:
#> stats::lm(formula = f, data = df1)
#> 
#> Residuals:
#>    Min     1Q Median     3Q    Max 
#> -26110 -21610  -6838  22721  56175 
#> 
#> Coefficients:
#>              Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  30750.03    2811.67  10.937   <2e-16 ***
#> linear_trend    19.18      17.73   1.082     0.28    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 23210 on 272 degrees of freedom
#> Multiple R-squared:  0.004288,	Adjusted R-squared:  0.0006274 
#> F-statistic: 1.171 on 1 and 272 DF,  p-value: 0.2801
#> 

# Plotting the residauls of the model
plot_res(md1)

# Adding monthly seasonal component
md2 <- trainLM(input = ny_gas,
               y = "y",
               trend = list(linear = TRUE),
               seasonal = "month")

plot_res(md2)

# Adding the first and seasonal lags

md3 <- trainLM(input = ny_gas,
               y = "y",
               trend = list(linear = TRUE),
               seasonal = "month",
               lags = c(1, 12))

plot_res(md3)

# Adding more lags and using stepwise regression for variable selection
md4 <- trainLM(input = ny_gas,
               y = "y",
               trend = list(linear = TRUE),
               seasonal = "month",
               lags = c(1:12),
               step = TRUE)

#> Start:  AIC=4468.62
#> y ~ month + linear_trend + lag_1 + lag_2 + lag_3 + lag_4 + lag_5 + 
#>     lag_6 + lag_7 + lag_8 + lag_9 + lag_10 + lag_11 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_8         1    1629786 5530867875 4466.7
#> - lag_2         1    2457031 5531695121 4466.7
#> - lag_7         1    2738340 5531976429 4466.8
#> - lag_4         1    4198603 5533436693 4466.8
#> - lag_11        1    7092093 5536330182 4467.0
#> - lag_6         1   10993921 5540232010 4467.1
#> - lag_12        1   18089458 5547327548 4467.5
#> - lag_3         1   19086109 5548324199 4467.5
#> <none>                       5529238089 4468.6
#> - lag_9         1   54085827 5583323917 4469.2
#> - lag_10        1   87144282 5616382372 4470.7
#> - lag_5         1  115893388 5645131477 4472.1
#> - linear_trend  1  135204542 5664442631 4473.0
#> - lag_1         1 1366603182 6895841271 4524.5
#> - month        11 2066473167 7595711257 4529.8
#> 
#> Step:  AIC=4466.7
#> y ~ month + linear_trend + lag_1 + lag_2 + lag_3 + lag_4 + lag_5 + 
#>     lag_6 + lag_7 + lag_9 + lag_10 + lag_11 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_7         1    1565211 5532433086 4464.8
#> - lag_2         1    2677078 5533544953 4464.8
#> - lag_4         1    3894285 5534762160 4464.9
#> - lag_11        1    7816040 5538683915 4465.1
#> - lag_6         1   11035084 5541902959 4465.2
#> - lag_12        1   17532972 5548400847 4465.5
#> - lag_3         1   17752665 5548620540 4465.5
#> <none>                       5530867875 4466.7
#> - lag_9         1   56469216 5587337091 4467.4
#> - lag_10        1   86157748 5617025624 4468.7
#> - lag_5         1  114837087 5645704962 4470.1
#> - linear_trend  1  138428739 5669296614 4471.2
#> - lag_1         1 1365058681 6895926556 4522.5
#> - month        11 2123312306 7654180181 4529.8
#> 
#> Step:  AIC=4464.77
#> y ~ month + linear_trend + lag_1 + lag_2 + lag_3 + lag_4 + lag_5 + 
#>     lag_6 + lag_9 + lag_10 + lag_11 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_2         1    3345461 5535778547 4462.9
#> - lag_4         1    4340404 5536773489 4463.0
#> - lag_11        1    7374695 5539807781 4463.1
#> - lag_6         1    9498322 5541931408 4463.2
#> - lag_3         1   17990825 5550423910 4463.6
#> - lag_12        1   19608730 5552041815 4463.7
#> <none>                       5532433086 4464.8
#> - lag_9         1   62222260 5594655345 4465.7
#> - lag_10        1   85044654 5617477740 4466.8
#> - lag_5         1  115824274 5648257360 4468.2
#> - linear_trend  1  137057229 5669490315 4469.2
#> - lag_1         1 1363493475 6895926561 4520.5
#> - month        11 2175652664 7708085750 4529.7
#> 
#> Step:  AIC=4462.93
#> y ~ month + linear_trend + lag_1 + lag_3 + lag_4 + lag_5 + lag_6 + 
#>     lag_9 + lag_10 + lag_11 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_4         1    4439915 5540218462 4461.1
#> - lag_11        1    6130765 5541909312 4461.2
#> - lag_6         1    9239131 5545017677 4461.4
#> - lag_12        1   22760158 5558538705 4462.0
#> - lag_3         1   29724197 5565502744 4462.3
#> <none>                       5535778547 4462.9
#> - lag_9         1   63617534 5599396081 4463.9
#> - lag_10        1   86905698 5622684244 4465.0
#> - lag_5         1  113901407 5649679954 4466.3
#> - linear_trend  1  140852389 5676630936 4467.5
#> - month        11 2196689070 7732467616 4528.5
#> - lag_1         1 1773296068 7309074615 4533.7
#> 
#> Step:  AIC=4461.14
#> y ~ month + linear_trend + lag_1 + lag_3 + lag_5 + lag_6 + lag_9 + 
#>     lag_10 + lag_11 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_11        1    6342758 5546561221 4459.4
#> - lag_6         1    8654181 5548872644 4459.6
#> - lag_12        1   22963693 5563182155 4460.2
#> - lag_3         1   25320141 5565538603 4460.3
#> <none>                       5540218462 4461.1
#> - lag_9         1   60161091 5600379553 4462.0
#> - lag_10        1   85187596 5625406058 4463.1
#> - linear_trend  1  137443532 5677661994 4465.6
#> - lag_5         1  158321622 5698540084 4466.5
#> - month        11 2209997778 7750216240 4527.1
#> - lag_1         1 1768882070 7309100532 4531.7
#> 
#> Step:  AIC=4459.44
#> y ~ month + linear_trend + lag_1 + lag_3 + lag_5 + lag_6 + lag_9 + 
#>     lag_10 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_6         1    7027742 5553588963 4457.8
#> - lag_3         1   24427780 5570989001 4458.6
#> <none>                       5546561221 4459.4
#> - lag_12        1   43962617 5590523837 4459.5
#> - lag_9         1   60198508 5606759729 4460.3
#> - lag_10        1  132668368 5679229588 4463.6
#> - linear_trend  1  143210965 5689772186 4464.1
#> - lag_5         1  155792917 5702354137 4464.7
#> - lag_1         1 1790153606 7336714827 4530.7
#> - month        11 2389580903 7936142124 4531.3
#> 
#> Step:  AIC=4457.77
#> y ~ month + linear_trend + lag_1 + lag_3 + lag_5 + lag_9 + lag_10 + 
#>     lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_3         1   26803252 5580392215 4457.0
#> - lag_12        1   41161451 5594750414 4457.7
#> <none>                       5553588963 4457.8
#> - lag_9         1   55355668 5608944631 4458.4
#> - lag_10        1  134403506 5687992468 4462.0
#> - linear_trend  1  159620884 5713209847 4463.2
#> - lag_5         1  166185873 5719774836 4463.5
#> - month        11 2386029111 7939618074 4529.4
#> - lag_1         1 1807306935 7360895898 4529.6
#> 
#> Step:  AIC=4457.04
#> y ~ month + linear_trend + lag_1 + lag_5 + lag_9 + lag_10 + lag_12
#> 
#>                Df  Sum of Sq        RSS    AIC
#> - lag_12        1   37449744 5617841959 4456.8
#> <none>                       5580392215 4457.0
#> - lag_9         1   65019043 5645411258 4458.1
#> - lag_10        1  135183409 5715575623 4461.3
#> - lag_5         1  141281752 5721673967 4461.6
#> - linear_trend  1  193153650 5773545865 4464.0
#> - lag_1         1 2136734094 7717126308 4540.0
#> - month        11 3152036395 8732428609 4552.4
#> 
#> Step:  AIC=4456.79
#> y ~ month + linear_trend + lag_1 + lag_5 + lag_9 + lag_10
#> 
#>                Df  Sum of Sq        RSS    AIC
#> <none>                       5.6178e+09 4456.8
#> - lag_9         1   60812664 5.6787e+09 4457.6
#> - lag_5         1  157502924 5.7753e+09 4462.0
#> - lag_10        1  170908172 5.7888e+09 4462.6
#> - linear_trend  1  227206510 5.8450e+09 4465.2
#> - lag_1         1 2389598793 8.0074e+09 4547.6
#> - month        11 6227113156 1.1845e+10 4630.2

summary(md4$model)

#> 
#> Call:
#> stats::lm(formula = y ~ month + linear_trend + lag_1 + lag_5 + 
#>     lag_9 + lag_10, data = df1)
#> 
#> Residuals:
#>    Min     1Q Median     3Q    Max 
#> -15595  -2180     -9   2013  15990 
#> 
#> Coefficients:
#>                Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)   3.365e+04  4.283e+03   7.857 1.24e-13 ***
#> monthFeb     -1.077e+04  1.961e+03  -5.491 9.99e-08 ***
#> monthMar     -1.276e+04  2.454e+03  -5.201 4.19e-07 ***
#> monthApr     -2.694e+04  2.936e+03  -9.173  < 2e-16 ***
#> monthMay     -2.764e+04  3.736e+03  -7.398 2.19e-12 ***
#> monthJun     -2.484e+04  4.527e+03  -5.488 1.01e-07 ***
#> monthJul     -2.334e+04  4.524e+03  -5.160 5.09e-07 ***
#> monthAug     -2.334e+04  4.279e+03  -5.455 1.20e-07 ***
#> monthSep     -2.636e+04  3.705e+03  -7.116 1.22e-11 ***
#> monthOct     -2.345e+04  3.445e+03  -6.805 7.69e-11 ***
#> monthNov     -1.346e+04  2.811e+03  -4.789 2.90e-06 ***
#> monthDec     -2.702e+03  2.071e+03  -1.305  0.19320    
#> linear_trend  1.524e+01  4.842e+00   3.148  0.00185 ** 
#> lag_1         5.344e-01  5.234e-02  10.208  < 2e-16 ***
#> lag_5        -1.390e-01  5.303e-02  -2.621  0.00932 ** 
#> lag_9        -1.023e-01  6.282e-02  -1.629  0.10470    
#> lag_10        1.758e-01  6.438e-02   2.730  0.00679 ** 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 4789 on 245 degrees of freedom
#> Multiple R-squared:  0.9607,	Adjusted R-squared:  0.9581 
#> F-statistic: 374.1 on 16 and 245 DF,  p-value: < 2.2e-16
#> 

plot_res(md4)